aya_obj
Module maps

Map struct and type bindings.

Structs

BtfMap

A BTF-defined map, most likely from a .maps section.

BtfMapDef

BTF definition of a map

LegacyMap

A map declared with legacy BPF map declaration style, most likely from a maps section.

bpf_map_def

Map definition in legacy BPF map declaration style

Enums

Map

Map data defined in maps or .maps sections

PinningError

The error type returned when failing to parse a PinningType

PinningType

The pinning type