aya_obj::obj
Struct Object

pub struct Object {
    pub endianness: Endianness,
    pub license: CString,
    pub kernel_version: Option<u32>,
    pub btf: Option<Btf>,
    pub btf_ext: Option<BtfExt>,
    pub maps: HashMap<String, Map>,
    pub programs: HashMap<String, Program>,
    pub functions: BTreeMap<(usize, u64), Function>,
    /* private fields */
}

The loaded object file representation

Fields
endianness: Endianness

The endianness

license: CString

Program license

kernel_version: Option<u32>

Kernel version

btf: Option<Btf>

Program BTF

btf_ext: Option<BtfExt>

Program BTF.ext

maps: HashMap<String, Map>

Referenced maps

programs: HashMap<String, Program>

A hash map of programs, using the program names parsed in ProgramSections as keys.

functions: BTreeMap<(usize, u64), Function>

Functions

Implementations

impl Object

pub fn fixup_and_sanitize_btf( &mut self, features: &BtfFeatures, ) -> Result<Option<&Btf>, BtfError>

Fixes up and sanitizes BTF data.

Mostly, it removes unsupported types and works around LLVM behaviours.

impl Object

pub fn relocate_btf( &mut self, target_btf: &Btf, ) -> Result<(), BtfRelocationError>

Relocates programs inside this object file with loaded BTF info.

impl Object

pub fn parse(data: &[u8]) -> Result<Object, ParseError>

Parses the binary data as an object file into an Object

pub fn patch_map_data( &mut self, globals: HashMap<&str, (&[u8], bool)>, ) -> Result<(), ParseError>

Patches map data

pub fn sanitize_functions(&mut self, features: &Features)

Sanitize BPF functions.

impl Object

pub fn relocate_maps<'a, I: Iterator<Item = (&'a str, RawFd, &'a Map)>>( &mut self, maps: I, text_sections: &HashSet<usize>, ) -> Result<(), EbpfRelocationError>

Relocates the map references

pub fn relocate_calls( &mut self, text_sections: &HashSet<usize>, ) -> Result<(), EbpfRelocationError>

Relocates function calls